Fix failing tests

+++ skainet-lang/skainet-lang-models/src/commonTest/kotlin/sk/ainet/lang/nn/compute/Rgb2GrayScaleMultiplyTest.kt
@@ -6,16 +6,16 @@ import sk.ainet.lang.model.compute.Rgb2GrayScaleMatMul
 import sk.ainet.lang.tensor.Shape
 import sk.ainet.lang.tensor.Tensor
 import sk.ainet.lang.tensor.dsl.tensor
-import sk.ainet.lang.types.FP32
+import sk.ainet.lang.types.FP16
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
 class Rgb2GrayScaleMultiplyTest {
 
-    private fun makeVoidInput1x1(rgb: Triple<Float, Float, Float>): Tensor<FP32, Float> {
+    private fun makeVoidInput1x1(rgb: Triple<Float, Float, Float>): Tensor<FP16, Float> {
         val (r, g, b) = rgb
-        return data<FP32, Float> {
-            tensor<FP32, Float> {
+        return data<FP16, Float> {
+            tensor<FP16, Float> {
                 shape(1, 3, 1, 1) {
                     fromArray(floatArrayOf(r, g, b))
                 }
@@ -27,13 +27,13 @@ class Rgb2GrayScaleMultiplyTest {
         r: Pair<Float, Float>,
         g: Pair<Float, Float>,
         b: Pair<Float, Float>
-    ): Tensor<FP32, Float> {
+    ): Tensor<FP16, Float> {
         // NCHW layout with W=2: [R0, R1, G0, G1, B0, B1]
         val (r0, r1) = r
         val (g0, g1) = g
         val (b0, b1) = b
-        return data<FP32, Float> {
-            tensor<FP32, Float> {
+        return data<FP16, Float> {
+            tensor<FP16, Float> {
                 shape(1, 3, 1, 2) {
                     fromArray(
                         floatArrayOf(
@@ -61,7 +61,7 @@ class Rgb2GrayScaleMultiplyTest {
 
         val output = module.forward(input)
         assertEquals(Shape(1, 1, 1, 2), output.shape, "Output shape must be (N,1,H,W)")
-        assertEquals(FP32::class, output.dtype)
+        assertEquals(FP16::class, output.dtype)
 
     }
